Minimal PoC for `pdm` as a dep manager
This provides a general lay out for a Python-Rust monorepo project using
`maturin` as a build backend.

```bash
python3.11 -m pip install pipx
pipx install --python python3.11 pdm
```

```
git clone https://github.com/TranSigma/gh-action-workbench
cd gh-action-workbench
pdm install --dev
eval $(pdm venv activate)
```
Once in this environment, the dev can develop as normal as if they were in any
normal venv *but* this is guaranteed to be consistent with the
[lock file](https://pdm.fming.dev/latest/usage/dependency/), and if using
`pdm update` frequently, be in sync with other developer's environments.

This is actually a very important point and goes with the other thing to note:
*there is no `pip` in this venv*. By default PDM does not create its venvs with
`pip`. This is to prevent a developer from monkeying with the packages outside
of the context of the project's listed dependencies. *If* intending to add
dependency to the project, then the correct use course of action would be to
use `pdm add` to actually apply it to the `pyproject.toml`—though, in this
case, the `-p path/to/local-dep` argument should be passed as the root
`pyproject.toml` is just a workspace config—if the intention is to add
new dependencies to a package.

If the dev wants to *experiment* with a package but doesn't want to modify the
`pyproject.toml`s, they can create a specific venv for this purpose using
something to the effect of `pdm venv create --name test-env --with-pip 3.11`
then use it for any testing they wish by passing the `--venv test-env` arg to
most commands.